requirements.txt
Flask>=3.0,<4
python-dotenv>=1.0,<2
SQLAlchemy>=2.0,<3
psycopg[binary]>=3.2,<4
PyMySQL>=1.1,<2
chromadb==0.5.5
sentence-transformers>=3.0,<4
openai>=1.43,<2
uvicorn>=0.30,<1
pydantic>=2.8,<3
jinja2>=3.1,<4
itsdangerous>=2.2,<3
click>=8.1,<9
werkzeug>=3.0,<4
requests>=2.32,<3
